pyproject.toml
[project]
name = "mlh-pe-hackathon"
version = "0.1.0"
requires-python = ">=3.13"
dependencies = [
    "faker>=33.0",
    "flask>=3.1",
    "peewee>=3.17",
    "psycopg2-binary>=2.9",
    "python-dotenv>=1.0",
    "pytest>=9.0.2",
    "pytest-cov>=7.1.0",
    "redis>=7.4.0",
    "gunicorn>=25.3.0",
    "psutil>=7.2.2",
    "prometheus-flask-exporter>=0.23.2",
]

